+++ nextgen/config/src/workspace/hasher_config.rs
@@ -1,3 +1,4 @@
+use crate::portable_path::GlobPath;
 use schematic::{derive_enum, Config, ConfigEnum};
 
 derive_enum!(
@@ -22,6 +23,10 @@ derive_enum!(
 pub struct HasherConfig {
     #[setting(default = 2500)]
     pub batch_size: u16,
+
+    pub ignore_patterns: Vec<GlobPath>,
+
+    pub ignore_missing_patterns: Vec<GlobPath>,
 
     pub optimization: HasherOptimization,