+++ packages/cli/CHANGELOG.md
@@ -17,7 +17,9 @@
 - Added support for `MOON_BASE` and `MOON_HEAD` environment variables.
   - Will be used when diffing across branches or commits.
   - Works for both `moon ci` and `moon run`.
-- Added `deno.bins` support in `.moon/toolchain.yml`.
+- Added `deno.bins` setting to `.moon/toolchain.yml`.
+- Added `hasher.ignorePatterns` and `hasher.ignoreMissingPatterns` settings to
+  `.moon/workspace.yml`.
 - Updated `moon ci` to include a summary of all failed actions.
 - Updated `moon run` to compare against the previous commit when running on the default branch and
   using `--remote`.